backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.45.0
python-dotenv==1.0.0
Pillow==10.0.1
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
httpx==0.25.2 